network restructure + deletions
