+++ 1_server/app.py
@@ -2,4 +2,4 @@
 from application import app
 # Runs in debug mode and allows all connections
 if __name__ == '__main__':
-	app.run(debug = True, port = 5004, host = "0.0.0.0")
+	app.run(debug = True, port = 5000, host = "0.0.0.0")
