+++ 2_name_generator/application/__init__.py
