+++ 2_name_generator/requirements.txt
