+++ 3_stats_generator/requirements.txt
