+++ 4_members_generator/app.py
@@ -1,5 +1,5 @@
 '''Main flask run file'''
-from flask_app import app
+from application import app
 # Runs in debug mode and allows all connections
 if __name__ == '__main__':
-	app.run(debug = True, host = '0.0.0.0')
+	app.run(debug = True, port = 5000, host = "0.0.0.0")
