--- name_generator/app.py
@@ -1,5 +1,0 @@
-'''Main flask run file'''
-from application import app
-# Runs in debug mode and allows all connections
-if __name__ == '__main__':
-	app.run(debug = True, port = 5002, host = "0.0.0.0")
