+++ nginx/nginx.conf
@@ -1,18 +1,18 @@
 events {}
 http {
 	server {
-        	listen 80;
-	        location /server {
-	        	proxy_pass http://server:5001;
+		listen 80;
+		location / {
+			proxy_pass http://server:5000;
 		}
 		location /service2 {
-			proxy_pass http://name_generator:5002;
+			proxy_pass http://name_generator:5000;
 		}
 		location /service3 {
-			proxy_pass http://stats_generator:5003;
+			proxy_pass http://stats_generator:5000;
 		}
 		location /service4 {
-			proxy_pass http://final_generation:5004;
+			proxy_pass http://members_generator:5000;
 		}
 	}
 }
